+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBiomeTagProvider.java
@@ -1403,8 +1403,10 @@ public final class WWBiomeTagProvider extends FrozenBiomeTagProvider {
 			.addOptional(WWWorldgen.ARID_FOREST);
 
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_STONE_PILE)
-			.add(Biomes.STONY_SHORE)
 			.addOptional(WWWorldgen.MAPLE_GROVE);
+
+		this.getOrCreateTagBuilder(WWBiomeTags.HAS_STONE_PILE_COMMON)
+			.add(Biomes.STONY_SHORE);
 
 		this.getOrCreateTagBuilder(WWBiomeTags.HAS_STONE_PILE_RARE)
 			.add(Biomes.SUNFLOWER_PLAINS)