+++ src/main/java/net/frozenblock/wilderwild/worldgen/modification/WWMiscGeneration.java
@@ -235,6 +235,10 @@ public final class WWMiscGeneration {
 							generationSettings.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, WWMiscPlaced.COARSE_DIRT_DISK_AND_PILE.getKey());
 						}
 
+						if (biomeSelectionContext.hasTag(WWBiomeTags.HAS_STONE_PILE_COMMON)) {
+							generationSettings.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, WWMiscPlaced.STONE_DISK_AND_PILE_COMMON.getKey());
+						}
+
 						if (biomeSelectionContext.hasTag(WWBiomeTags.HAS_STONE_PILE)) {
 							generationSettings.addFeature(GenerationStep.Decoration.UNDERGROUND_ORES, WWMiscPlaced.STONE_DISK_AND_PILE.getKey());
 						}